% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcpredictions.R
\name{as_list_format}
\alias{as_list_format}
\title{A quick replacement to \code{\link[runjags:list.format]{runjags::list.format()}} that does nothing if the argument is already a list.}
\usage{
as_list_format(data, checkvalid = TRUE)
}
\arguments{
\item{data}{Same as \code{\link[runjags:list.format]{runjags::list.format()}}. Typically found in the \code{data} slot of runjags object.}

\item{checkvalid}{See \code{\link[runjags:list.format]{runjags::list.format()}}.}
}
\description{
A quick replacement to \code{\link[runjags:list.format]{runjags::list.format()}} that does nothing if the argument is already a list.
}
