% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks_of_rs_products.R
\name{brick_gpp}
\alias{brick_gpp}
\alias{brick_pg}
\alias{brick_ssoil}
\alias{brick_fmc}
\alias{brick_woodycover}
\title{Extract 8 Day Frequency Bricks for GPP}
\usage{
brick_gpp(spobj, years)

brick_pg(spobj, years)

brick_ssoil(spobj, years)

brick_fmc(spobj, years)

brick_woodycover(spobj, years)
}
\arguments{
\item{spobj}{Spatial* object that informs extents of the raster to extract}

\item{years}{Years of GPP to extract}
}
\value{
A raster brick with extent given by \code{extent(spobj)}
}
\description{
Extract 8 Day Frequency Bricks for GPP
}
\section{Functions}{
\itemize{
\item \code{brick_pg}: Extract brick of pg values

\item \code{brick_ssoil}: Extract brick of soil moisture values

\item \code{brick_fmc}: Extract brick of 8 day fmc values

\item \code{brick_woodycover}: Extract a raster brick for Woody Cover data
}}

