% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_filename_list.R
\name{build_filename_list}
\alias{build_filename_list}
\title{Create list of Filenames}
\usage{
build_filename_list(
  server,
  directory = "",
  prefix = "",
  middle = "",
  suffix = "",
  type_extension = "",
  dirsep = "/",
  namesep = "."
)
}
\arguments{
\item{server}{server url}

\item{directory}{directory of data within server}

\item{prefix}{start of filename}

\item{middle}{middle of filename}

\item{suffix}{end of filename}

\item{type_extension}{The character string representing file type}

\item{dirsep}{Character used to combine server and directory}

\item{namesep}{Characer used to combine parts of the filenames}
}
\description{
Create list of Filenames
}
\details{
server and directory are concatenated using dirsep. prefix, middle, suffix and type_extension are concatenated using namesep separator.
Uses \code{paste}.
}
\examples{
server <- "http://dapds00.nci.org.au/thredds/dodsC/ub8/au/OzWALD" #http://dapds00.nci.org.au/thredds/catalog/ub8/au/OzWALD/catalog.html
directory <- "8day/GPP"
prefix <- "OzWALD"
middle <- "GPP"
suffix <- 2000:2018
type_extension <- "nc"
build_filename_list("http://dapds00.nci.org.au/thredds/dodsC/ub8/au/OzWALD",
      "8day/GPP",
      "OzWALD",
      "GPP",
      2000:2018,
      "nc")
}
