% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_into_tiles.R
\name{divide_into_tiles}
\alias{divide_into_tiles}
\alias{make_grid_buffer}
\title{Divide points into tiles}
\usage{
divide_into_tiles(points, cellsize = 1, buffer = 0.01)

make_grid_buffer(sfobj, cellsize = 0.1, overlap = 0.01)
}
\arguments{
\item{points}{is an sf object (rather like a dataframe) of point geometries}

\item{cellsize}{is the size of a single cell}

\item{buffer}{is the spatial distance required around each point
(each point must have a tile that contains buffer distance region around it)}

\item{sfobj}{An sf object to cover with a grid. Could also be a bbox object.}

\item{overlap}{is the amount of overlap of each cell}
}
\value{
A list. Each element contains a slot 'tile' which is an sfc_POLYGON object,
and a slot 'pts', which is a an sf object of points covered by the tile with sufficient buffer

An sf multipolygon object made of squares that overlap by buffer, and cover the region bbox expanded by buffer.
}
\description{
Divide points into tiles
}
\section{Functions}{
\itemize{
\item \code{make_grid_buffer}: Creates a full grid of overlapping polygons that covers sfobj.
}}

