% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epbc_birds.R
\name{download_epbc}
\alias{download_epbc}
\title{Downloads the EPBC Lists of Species}
\usage{
download_epbc()
}
\arguments{
\item{destfile}{The file to save the epbc lists to.}
}
\value{
A data frame of the listed species with columns for Scientific Name, Common Name, Fauna Type, Conservation Status, and a few others columns present on the website.
The cell values have been processed as little as possible. So, for example, a 'Common Name' for Pseudocheirus occidentalis is Western Ringtail Possum, Ngwayir, Womp, Woder, Ngoor, Ngoolangit.
}
\description{
Converts the lists shown at https://www.environment.gov.au/cgi-bin/sprat/public/publicthreatenedlist.pl to a table.
}
\details{
Requires the packages rvest and xml2. Worked on March 23, 2021.
}
\examples{
constatuses <- download_epbc()
destfile = paste0("./private/data/raw/EPBClist_", Sys.Date(), ".tsv")
write.table(constatuses, destfile, sep = "\t")
}
