% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_residuals.R
\name{ds_detection_residuals.fit}
\alias{ds_detection_residuals.fit}
\alias{ds_occupancy_residuals.fit}
\title{Dunn-Smyth Residuals for Detection and Occupancy for Fitted Models}
\usage{
ds_detection_residuals.fit(fit, type = "median", seed = NULL)

ds_occupancy_residuals.fit(
  fit,
  type = "median",
  seed = NULL,
  conditionalLV = TRUE
)
}
\arguments{
\item{fit}{A fitted occupancy-detection model.}

\item{type}{The type of point estimate to use for parameter estimates. See \code{\link{get_theta}}}

\item{seed}{A seed to fix randomness of Dunn-Smyth residual jitter.}
}
\value{
A matrix, each row is a ModelSite and each column is a species.
Detection residuals are only computed for species and sites that have at least one detection. Other values are NA.

A matrix, each row is a ModelSite and each column is a species.
}
\description{
Dunn-Smyth Residuals for Detection and Occupancy for Fitted Models
}
\section{Functions}{
\itemize{
\item \code{ds_detection_residuals.fit}: Given a fitted occupancy detection model (variable names must match). Computes Dunn-Smyth residuals for detection, marginalising the latent variables.

\item \code{ds_occupancy_residuals.fit}: Given a fitted occupancy detection model (variable names must match).
Computes Dunn-Smyth residuals for occupancy, marginalising the latent variables.
}}

\examples{
fit <- readRDS("./tmpdata/7_2_10_clim_year.rds")
detection_resids <- ds_detection_residuals.fit(fit, type = "median")
occupancy_resids <- ds_occupancy_residuals.fit(fit, type = "median", conditionalLV = FALSE)
}
\references{
D. I. Warton, J. Stoklosa, G. Guillera-Arroita, D. I. MacKenzie, and A. H. Welsh,
"Graphical diagnostics for occupancy models with imperfect detection," Methods in Ecology and Evolution, vol. 8, no. 4, pp. 408-419, 2017, doi: 10.1111/2041-210X.12761.
}
