% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_biodiversity_multisite.R
\name{expectedspeciesnum.theta}
\alias{expectedspeciesnum.theta}
\title{Expected Biodiversity Across Multiple Sites}
\usage{
expectedspeciesnum.theta(Xocc, u.b, lv.coef = NULL, LVvals = NULL)
}
\description{
Assuming independent patches, estimates the posterior expected number of species in across all patches.
}
