% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_detectionoccupancy.R
\name{extend.fit}
\alias{extend.fit}
\title{Extend MCMC chains of fitted model}
\usage{
extend.fit(fit, ..., filename = NULL)
}
\arguments{
\item{fit}{Model fitted by run.detectionoccupancy}

\item{...}{arguments passed to extent.jags}

\item{filename}{File to save extended chains (with modelling information) to.}
}
\description{
Extend MCMC chains of fitted model
}
