% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_poly.R
\name{extent_poly}
\alias{extent_poly}
\title{Helper function to get the extent polygon of a raster}
\usage{
extent_poly(x)
}
\arguments{
\item{x}{A raster* object from the \code{raster} package}
}
\description{
Helper function to get the extent polygon of a raster
}
\examples{
ras <- raster::raster("./inst/demoBGGWmap.tif")
roi <- extent_poly(ras)
}
