% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_brick.R
\name{fetch_brick_Albers}
\alias{fetch_brick_Albers}
\title{Extract a brick of Australian Albers Tiles.}
\usage{
fetch_brick_Albers(
  spobj,
  years,
  get_tile_filenames = get_bggwtile_filenames,
  tilereader = bggwtilereader
)
}
\arguments{
\item{spobj}{Spatial* or sf object that informs extents of the raster to extract}

\item{years}{Years of data to extract}

\item{get_tile_filenames}{A function with arguments (tilecode, years, ...).
For a given tilecode and years it must return a list of filenames. Each filename is for the tile for each year provided.}

\item{tilereader}{A function that accepts a single argument, 'filename', and returns a raster object for that file.}
}
\value{
A raster brick with extent equal or larger than \code{extent(spobj)}, snapped to the cells of the raster data.
The projection of the returned raster is EPSG:3577, which is GDA94.
Extent of the returned value is a rectangle, and pixel values outside \code{spobj} are included.
}
\description{
Uses the files stored at the given location, and assumes the tiles are saved as EPSG:3577, which is GDA94.
}
\examples{

}
