% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woody_vals_buffer.R
\name{fetch_woody_cover_meanbuffer}
\alias{fetch_woody_cover_meanbuffer}
\title{Extract mean woody cover within buffers for a set of points}
\usage{
fetch_woody_cover_meanbuffer(pts, years, buffers)
}
\arguments{
\item{pts}{is a set of sf points within roi}

\item{years}{Years of data to acquire}

\item{buffers}{Multiple buffers given in METRES.}

\item{roi}{is an sf object.}
}
\value{
A list corresponding to each buffer distance. In each element each row is a point, each column is a date
}
\description{
Extract mean woody cover within buffers for a set of points
}
\examples{
locations <- read.csv("./private/data/clean/site_locations_cleaned.csv")[1:2, ]
pts <- sf::st_as_sf(locations, coords = c("MeanLON", "MeanLAT"), crs = 4326)[1:2, ]
years <- 2000:2001
buffers <- c(100, 500)
global_muffle_discarddatumwarn()
fetch_woody_cover_meanbuffer(pts, years, buffers)
}
