% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal_bylayer.R
\name{focal_bylayer}
\alias{focal_bylayer}
\title{Apply raster's focal function by layer}
\usage{
focal_bylayer(
  x,
  w,
  fun,
  na.rm = FALSE,
  pad = FALSE,
  padValue = NA,
  NAonly = FALSE,
  cl = NULL,
  ...
)
}
\arguments{
\item{x}{A RasterBrick or RasterStack}

\item{w}{Focal weights as required by \code{focal}}

\item{fun}{Function to apply to value in focal region}

\item{na.rm}{If TRUE NA values removed from computations.}

\item{pad}{Passed to \code{\link[raster:focal]{raster::focal()}}. If TRUE, additional 'virtual' rows and columns are padded to x such that there are no edge effects. This can be useful when a function needs to have access to the central cell of the filter.}

\item{padValue}{Passed to \code{\link[raster:focal]{raster::focal()}}. The value of the cells of the padded rows and columns.}

\item{NAonly}{Passed to \code{\link[raster:focal]{raster::focal()}}. If TRUE, only cell values that are NA are replaced with the computed focal values.}

\item{cl}{a cluster for parallel computation created par \pkg{parallel}. If NULL then no parallel computations will occur.}
}
\description{
Apply's raster's \code{focal} function to layers in a brick or stack. See help for \code{focal} for more information on paramater behaviour.
}
