% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlbertDrills.R
\name{getsite_fromonetwothreeposts}
\alias{getsite_fromonetwothreeposts}
\title{Map Albert's Remotely Sensed Data Drills to Three Post Sites}
\usage{
getsite_fromonetwothreeposts(
  valabbrv,
  timecode,
  threepostsfile = "./AlbertData/threeposts_wgs84.csv"
)
}
\arguments{
\item{valabbrv}{Abbreviation of value name used in Albert's file naming convention.}

\item{timecode}{The time code (typically year) that is used in Albert's files.}

\item{threepostsfile}{The csv file of post locations used by Albert.
This file contains the site and study information in the same order as the remote sensed values.}
}
\value{
A tibble with columns Study SiteCode post, LON, LAT,
and data columns named by valabbrv_timecode_radius (e.g. maxBS_2000_1000)
}
\description{
Map Albert's Remotely Sensed Data Drills to Three Post Sites
}
\examples{
out <- getsite_fromonetwothreeposts("maxBS", 2001, "./AlbertData/threeposts_wgs84.csv")
}
