% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muffle_datumwarn.R
\name{global_muffle_discarddatumwarn}
\alias{global_muffle_discarddatumwarn}
\title{Muffle Discarded datum Unknown warnings}
\usage{
global_muffle_discarddatumwarn()
}
\description{
Globally muffles all warnings that contain 'Discarded datum Unknown'
}
