% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpp_vals.R
\name{gpp_vals}
\alias{gpp_vals}
\title{Extract GPP values at points}
\usage{
gpp_vals(roi, pts, years)
}
\arguments{
\item{roi}{is an sf object.}

\item{pts}{is a set of sf points within roi}

\item{years}{Years of data to acquire}
}
\value{
each row is a point, each column is a date
}
\description{
Extract GPP values at points
}
