% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_analysis.R
\name{gvifs_lm}
\alias{gvifs_lm}
\title{Compute VIF using CAR package}
\usage{
gvifs_lm(Xocc, yXobs, species, ModelSite, OccFmla = "~ 1", ObsFmla = "~ 1")
}
\arguments{
\item{Xocc}{A dataframe of covariates related to occupancy. One row per ModelSite.
Must also include the ModelSiteVars columns to uniquely specify ModelSite.}

\item{yXobs}{A dataframe of species observations (1 or 0) and covariates related to observations. One row per visit.
Each column is either a covariate or a species.
Must also include the ModelSiteVars columns to uniquely specify ModelSite.}

\item{species}{A list of species names (corresponding to columns of yXobs) to model.}

\item{ModelSite}{A list of column names in y, Xocc and Xobs that uniquely specify the ModelSite. Can be simply a ModelSite index}

\item{OccFmla}{A formula specifying the occupancy model in terms of the covariates in Xocc.}

\item{ObsFmla}{A formula specifying the model for detection in terms of the covariates in Xobs.}
}
\description{
Compute VIF using CAR package
}
\examples{
inputdata <- readRDS("./private/data/clean/7_2_10_input_data.rds")

OccFmla <- reformulate(paste0("\`",setdiff(names(inputdata$insample$Xocc), c("ModelSiteID", "SiteCode", "SurveySiteId", "StudyId", "holdout", "gc", "AnnTempRange")), "\`"))
ObsFmla <- "~ 1 + MeanWind + MeanTime + MeanClouds + MeanTemp"
vifs <- gvifs_lm(inputdata$insample$Xocc,
         inputdata$insample$yXobs,
         species = inputdata$species,
         ModelSite = "ModelSiteID",
         OccFmla = OccFmla,
         ObsFmla = ObsFmla
         )

vifs[[1]] \%>\%
  tibble::enframe(name = "var", value = "vif") \%>\%
  arrange(vif) \%>\%
  ggplot() +
  geom_point(aes(y = var, x = vif)) +
  geom_vline(xintercept = 30)
}
