% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapplyquery.R
\name{lapplyquery}
\alias{lapplyquery}
\title{SQL Helper Function: Apply Query multiple times, substituting the '?' in the query}
\usage{
lapplyquery(qry, vallist)
}
\value{
A data frame of all results bound together (bound using rbind)
}
\description{
Uses dbBind to fetch queries using wildcard values given in vallist.
}
