% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LV_covar_plots.R
\name{plot_LVvscovar.fit}
\alias{plot_LVvscovar.fit}
\alias{plotdf_LVvscovar.fit}
\title{Plot Estimated Latent Variables Against Covariates}
\usage{
plot_LVvscovar.fit(
  fit,
  esttype = "median",
  theta = NULL,
  covar,
  facetvars = NULL,
  cuts = 3,
  aggregatefcn = mean
)

plotdf_LVvscovar.fit(
  fit,
  esttype = "median",
  theta = NULL,
  covar,
  aggregatefcn = mean
)
}
\arguments{
\item{fit}{A fitted runjags object.}

\item{esttype}{When fit is non-NULL, then esttype is used to extract parameter values}

\item{theta}{a vector of model parameters with BUGS names}

\item{covar}{a dataframe of covariate values. It must have a column labelled 'ModelSite' that gives the ModelSite of covariate value.}

\item{aggregatefcn}{Rows in covar with duplicate ModelSite values are aggregrated using this function}
}
\description{
Plot Estimated Latent Variables Against Covariates
}
\section{Functions}{
\itemize{
\item \code{plot_LVvscovar.fit}: Plot estimated latent variable values against covariate values for occupancy

\item \code{plotdf_LVvscovar.fit}: Prepare plotting data frames for latent variable values against covariate values for occupancy
}}

