% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_quality_stats.R
\name{modelqualstats}
\alias{modelqualstats}
\title{Compute Model Quality Information}
\usage{
modelqualstats(fit, holdoutXocc, holdoutyXobs, ModelSite, cl)
}
\arguments{
\item{fit}{a model fitted using run.detectionoccupancy}

\item{holdoutXocc}{A data frame of occupancy predictors for sites.}

\item{holdoutyXobs}{A data frame of detection information and species detections}

\item{ModelSite}{Column names used to match the holdoutyXobs to sites in holdoutXocc}

\item{cl}{A cluster created by parallel::makeCluster}
}
\description{
Compute Model Quality Information
}
