% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordfactby.R
\name{ordfactby}
\alias{ordfactby}
\title{Turn a column into an ordered factor}
\usage{
ordfactby(data, column, by, ...)
}
\arguments{
\item{data}{a data frame or tibble}

\item{column}{the column to convert to an ordered factor}

\item{by}{the column used to order \code{column}}

\item{...}{expressions passed to \code{dplyr::filter}. Only rows satisfying these expressions are used to order \code{column},}
}
\value{
The dataframe with \code{column} converted to an ordered factor that is ordered according to \code{by}
}
\description{
Converts a column into an ordered factor.
}
\details{
\code{column}, \code{by}, and \code{...} use tidyselect (e.g. column names do not need to be quoted)
}
\examples{
library(tidyr)
mtcars$Model <- rownames(mtcars)
ordfactby(mtcars, Model, gear)  #turns column Model into an ordered factor ordered by gear
}
