% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcspeciesprobs_singlesitetheta.R
\name{pdetection_occupied.ModelSite.theta}
\alias{pdetection_occupied.ModelSite.theta}
\alias{poccupy.ModelSite.theta}
\title{Probability of detection and occupation for a single ModelSite and parameter set}
\usage{
pdetection_occupied.ModelSite.theta(Xobs, v.b)

poccupy.ModelSite.theta(Xocc, u.b, lv.coef = NULL, LVvals = NULL)
}
\arguments{
\item{Xobs}{A matrix of detection covariates, each row is a visit.}

\item{v.b}{Covariate loadings. Each row is a species, each column a detection covariate}

\item{Xocc}{A matrix of occupancy covariates. Must have a single row. Columns correspond to covariates.}

\item{u.b}{Covariate loadings for occupancy. Each row is a species, each column an occupancy covariate.}

\item{lv.coef}{Loadings for the latent variables. Each row is a species, each column corresponds to a LV.}

\item{LVvals}{A matrix of LV values. Each column corresponds to a LV. To condition on specific LV values, provide a matrix of row 1.}
}
\value{
A matrix of detection probabilities for each row of Xobs and each species.
Rows in returned value correspond to rows in Xobs, columns correspond to species.

A matrix of occupancy probabilities. Each row corresponds to a row of LVvals, each column to a species.
}
\description{
Probability of detection and occupation for a single ModelSite and parameter set
}
\section{Functions}{
\itemize{
\item \code{poccupy.ModelSite.theta}: The probability of occupation for given LV values.
}}

