% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_residuals_plots.R
\name{plot_residuals.residual}
\alias{plot_residuals.residual}
\alias{facet_species_covariate}
\alias{facet_covariate}
\alias{plot_residuals_detection.fit}
\alias{plot_residuals_occupancy.fit}
\title{Plotting of Dunn-Smyth Residuals}
\usage{
plot_residuals.residual(
  residuals,
  covar,
  plotfunction = facet_species_covariate,
  aggregatefcn = mean,
  ...
)

facet_species_covariate(data, ...)

facet_covariate(data, ...)

plot_residuals_detection.fit(
  fit,
  detectionresiduals = NULL,
  varidx = NULL,
  esttype = NULL,
  aggregatefcn = mean,
  ...
)

plot_residuals_occupancy.fit(
  fit,
  occupancyresidual = NULL,
  varidx = NULL,
  esttype = NULL,
  conditionalLV = TRUE,
  aggregatefcn = mean,
  ...
)
}
\arguments{
\item{residuals}{is a table returned by \code{ds_occupancy_residuals.fit} or \code{ds_detection_residuals.fit}. Each row is a unique ModelSite.
It must have a column named 'ModelSite', all other columns are assumed contain residuals, and the column name corresponds to species names.}

\item{covar}{Is a table of covariate values, it must have a column labelled 'ModelSite' that gives the ModelSite of covariate value.
Rows with duplicated ModelSite values are averaged, which is in keeping with Warton et al for detection residuals.}

\item{plotfunction}{A plotting method to use. Default is \code{facet_species_covariate}.}

\item{...}{Extra arguments to pass. Currently ignored (no extra arguments accepted).}

\item{data}{Input tibble. Columns of Species, Residual, Covariate and CovariateValue}

\item{fit}{The fitted runjags object.}

\item{detectionresiduals}{Optional. A tibble of already calculated detection residuals.
Must have column names identical to output of \code{ds_detection_residuals.fit}.
If not supplied then detection residuals are computed from \code{fit} using \code{ds_detection_residuals.fit}.}

\item{varidx}{The index of the covariate to plot. If NULL, then all occupancy covariates will be plotted}

\item{esttype}{The point estimate extracted from fit. Passed to \code{ds_detection_residuals.fit} as argument \code{type}.}

\item{occupancyresiduals}{Optional. A tibble of already calculated occupancy residuals.
Must have column names identical to output of \code{ds_detection_residuals.fit}.
If not supplied then detection residuals are computed from \code{fit} using \code{ds_detection_residuals.fit}.}
}
\value{
A ggplot object. Data is in the \code{data} slot.

A ggplot object.

A ggplot object.

A ggplot object. Data is saved in the \code{data} slot.

A ggplot object. Data is saved in the \code{data} slot.
}
\description{
Plotting of Dunn-Smyth Residuals
}
\details{
The inclusion of a smoothed fit for easier interpretability follows suggestions by Warton et al (2017)
}
\section{Functions}{
\itemize{
\item \code{plot_residuals.residual}: For table of provided residuals and covariate values, makes residual plots.
Residual and covariate values must be provided with the ModelSite.

\item \code{facet_species_covariate}: A function that prepares plot of residuals, one facet for each species and covariate

\item \code{facet_covariate}: A function that prepares plot of residuals, one facet for each covariate. Species ignored.

\item \code{plot_residuals_detection.fit}: Prepares tibbles and plots of residuals for covariates that are part of a fitted object

\item \code{plot_residuals_occupancy.fit}: Prepares tibbles and plots of occupancy residuals for covariates that are part of a fitted object
}}

\examples{
fit <- readRDS("./tmpdata/7_1_mcmcchain_20200424.rds")
fit <- runjags::add.summary(fit)
source("./R/calcpredictions.R")
source("./R/DS_residuals.R")
fit$data <- as_list_format(fit$data)
detection_resids <- ds_detection_residuals.fit(fit, type = "median", conditionalLV = FALSE)

# Plot Detection Residual 
plt <- plot_residuals_detection.fit(fit, detection_resids, varidx = 2, aggregatefcn = max, conditionalLV = FALSE)
plt
plt + coord_cartesian(ylim = c(-1, +1))
plot_residuals_detection.fit(fit, varidx = 2, esttype = "median", conditionalLV = FALSE)
plot_residuals_detection.fit(fit, varidx = c(2, 3), esttype = "median",
  plotfunction = facet_covariate, conditionalLV = FALSE) + 
  coord_cartesian(ylim = c(-1, +1))

plot_residuals_occupancy.fit(fit, varidx = 3, esttype = "median", conditionalLV = FALSE)
plot_residuals_occupancy.fit(fit, varidx = c(2, 3), esttype = "median",
 plotfunction = facet_covariate, conditionalLV = FALSE)

# Residuals against an unincluded covariate:
source("./scripts/7_1_import_site_observations.R")
covar <- occ_covariates[ , "ms_cover", drop = FALSE] \%>\%
  tibble::rowid_to_column(var = "ModelSite")
residuals <- ds_occupancy_residuals.fit(fit, type = "median", conditionalLV = FALSE)
pltobj <- plot_residuals.residual(residuals, covar, 
                 plotfunction = facet_species_covariate)
pltobj + scale_y_continuous(name = "Occupancy Residuals") 
pltobj + scale_y_continuous(name = "Occupancy Residuals") + coord_cartesian(ylim = c(-1, +1))

## Example using 7_2_1 data
data_7_2_1 <- readRDS("./private/data/clean/7_2_1_input_data.rds")
fitp <- readRDS("./tmpdata/7_2_2_detectiononly_smallAmodel_run_20200529.rds")
fit <- results.jags("./runjagsfiles_13", read.monitor = c("LV", fitp$monitor))
fit <- add.summary(fit)
covar <- data_7_2_1$plotsmerged_detection \%>\%
  dplyr::select(-all_of(data_7_2_1$detection_data_specieslist)) \%>\%
  dplyr::select_if(is.numeric) \%>\%
  rename(ModelSite = ModelSiteID)
plot_residuals_detection.fit(fit, varidx = 2, esttype = "median", conditionalLV = TRUE) + coord_cartesian(ylim = c(-1, 1))
plt <- plot_LVvscovar.fit(fit, esttype = "median", covar = covar, aggregatefcn = min)
plt + coord_cartesian(ylim = c(-1, 1))
}
\references{
D. I. Warton, J. Stoklosa, G. Guillera-Arroita, D. I. MacKenzie, and A. H. Welsh,
"Graphical diagnostics for occupancy models with imperfect detection," Methods in Ecology and Evolution, vol. 8, no. 4, pp. 408-419, 2017, doi: 10.1111/2041-210X.12761.
}
