% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poccupy.R
\name{poccupy.ModelSite}
\alias{poccupy.ModelSite}
\title{Probability of each species occupancy marginal across other species, from the full posterior distribution.}
\usage{
poccupy.ModelSite(Xocc, u.b_arr, lv.coef_arr = NULL, LVvals = NULL)
}
\arguments{
\item{u.b_arr}{A 3-array of occupancy covariates. Each row is a species, each column is a covariate, and layer is a draw from the posterior.}

\item{lv.coef_arr}{A 3-array of LV loadings. If \code{NULL} then model is assumed to not have latent variables.}

\item{LVvals}{If a matrix then LVvals are considered simulated and for each possible set of parameters,
the probabilities are marginalised of the (prior) LV distribution.
If \code{LVvals} is an array with 3 dimensions, then the third dimension corresponds to the posterior distribution draw (same as u.b_arr).
In this situation the array must have only one row: the row corresponding the model site.}
}
\description{
Probability of each species occupancy marginal across other species, from the full posterior distribution.
}
\details{
Compute the probability of each species individually (non-joint) using the full posterior distribution for a single ModelSite.
}
