% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numspec_fulldistribution.R
\name{predsumspeciesRV}
\alias{predsumspeciesRV}
\alias{predsumspeciesRV_newdata}
\alias{sumspeciesRV_raw}
\title{Compute predicted distributions of number species detected}
\usage{
predsumspeciesRV(
  fit,
  chains = NULL,
  UseFittedLV = TRUE,
  nLVsim = 100,
  type = "marginal",
  cl = NULL
)

predsumspeciesRV_newdata(
  fit,
  Xocc,
  Xobs = NULL,
  ModelSiteVars = NULL,
  chains = NULL,
  nLVsim = 100,
  type = "marginal",
  cl = NULL
)

sumspeciesRV_raw(
  Xocc,
  Xobs = NULL,
  ModelSite = NULL,
  numspecies,
  nlv,
  draws,
  useLVindraws = TRUE,
  nLVsim = NULL,
  cl = NULL
)
}
\arguments{
\item{fit}{A runjags fitted object}

\item{chains}{The chains of MCMC to use. Default is all chains.}

\item{UseFittedLV}{If TRUE the fitted LV variables are used, if false then 100 LV values are simulated.}

\item{type}{If "draws" then a list of distributions for each site and draw is returned
If "marginal" then a single distribution for each site is returned and this distribution is marginalised over the draws.}

\item{cl}{A cluster object created by parallel::makeCluster. If NULL no cluster is used.}

\item{Xocc}{A matrix of occupancy covariates. Must have a single row. Columns correspond to covariates.}

\item{Xobs}{A matrix of detection covariates, each row is a visit. If NULL then expected number of species in occupation is returned}

\item{theta}{A vector of model parameters, labelled according to the BUGS labelling convention seen in runjags}

\item{LVvals}{A matrix of LV values. Each column corresponds to a LV. To condition on specific LV values, provide a matrix of row 1.}
}
\description{
Uses \link{discreteRV} to compute distribution of number of species
}
\section{Functions}{
\itemize{
\item \code{predsumspeciesRV_newdata}: Distribution of species numbers for new ModelSite

\item \code{sumspeciesRV_raw}: For raw model information. Returns distributions per draw.
}}

\examples{
fit <- readRDS("./tmpdata/7_3_02_clim_someclimate_year_woody500m_msnm_det1stO.rds")
indata <- readRDS("./private/data/clean/7_2_10_input_data.rds")
det_dist_2060 <- predsumspeciesRV_newdata(fit,
                         Xocc = indata$holdoutdata$Xocc[indata$holdoutdata$Xocc$ModelSiteID == 2060, , drop = FALSE],
                         Xobs = indata$holdoutdata$yXobs[indata$holdoutdata$yXobs$ModelSiteID == 2060, , drop = FALSE],
                         ModelSiteVars = "ModelSiteID")
}
