% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmprasterdir.R
\name{rastertmpdir_movetonew}
\alias{rastertmpdir_movetonew}
\alias{rastertmpdir_movetoorig}
\alias{rastertmpdir_removenew}
\title{Functions for using and then deleting new temporary directories}
\usage{
rastertmpdir_movetonew(newdir = NULL)

rastertmpdir_movetoorig(origdir)

rastertmpdir_removenew(newdir)
}
\arguments{
\item{newdir}{New directory name within the current raster temporary direcotory.
If NULL it will be generated on creation.}

\item{origdir}{Original temporary raster directory}
}
\description{
The raster package does large computations by saving data temporarily to the computer harddisk.
However, the package does not know when to delete these files so they are left there.
For large data computations the hard drive can quickly fill with intermediate data files.
Any data that you want to save in the temporary directory must moved elsewhere before deleting the
temporary directory.
}
\section{Functions}{
\itemize{
\item \code{rastertmpdir_movetoorig}: Switches raster temporary directory back to original directory.

\item \code{rastertmpdir_removenew}: Deletes the temporary directory.
}}

\examples{
dirs <- rastertmpdir_movetonew()
rastertmpdir_movetoorig(dirs$orig)
# shift data to save
rastertmpdir_removenew(dirs$new)
}
