% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_detectionoccupancy.R
\name{run.detectionoccupancy}
\alias{run.detectionoccupancy}
\alias{prep.data}
\alias{prep_new_data}
\alias{defaultinitsfunction}
\title{Fit detection occupancy models using runjags.}
\usage{
run.detectionoccupancy(
  Xocc,
  yXobs,
  species,
  ModelSite,
  OccFmla = "~ 1",
  ObsFmla = "~ 1",
  nlv = 2,
  initsfunction = defaultinitsfunction,
  MCMCparams = list(n.chains = 1, adapt = 2000, burnin = 25000, sample = 1000, thin =
    30),
  filename = NULL
)

prep.data(Xocc, yXobs, ModelSite, species, nlv, XoccProcess, XobsProcess)

prep_new_data(fit, Xocc, yXobs, ModelSite)

defaultinitsfunction(chain, indata, ...)
}
\arguments{
\item{Xocc}{A dataframe of covariates related to occupancy. One row per ModelSite.
Must also include the ModelSiteVars columns to uniquely specify ModelSite.}

\item{yXobs}{A dataframe of species observations (1 or 0) and covariates related to observations. One row per visit.
Each column is either a covariate or a species.
Must also include the ModelSiteVars columns to uniquely specify ModelSite.}

\item{species}{A list of species names (corresponding to columns of yXobs) to model.}

\item{ModelSite}{A list of column names in y, Xocc and Xobs that uniquely specify the ModelSite. Can be simply a ModelSite index}

\item{OccFmla}{A formula specifying the occupancy model in terms of the covariates in Xocc.}

\item{ObsFmla}{A formula specifying the model for detection in terms of the covariates in Xobs.}

\item{nlv}{The number of latent variables}

\item{MCMCparams}{A named list containing values for n.chains, adapt, burnin, sample and thin (to pass to run.jags).
Also include keep.jags.files to specify the directory that JAGS data will be saved.}

\item{filename}{If non-null the runjags object (with some extra information) is saved to filename as an RDS.}

\item{XoccProcess}{An object create by prep.designmatprocess for the occupancy covariates}

\item{XobsProcess}{An object create by prep.designmatprocess for the observation covariates}

\item{chain}{Integer. Index of the chain.}

\item{indata}{A list of data that is typically passed to \code{\link[runjags:run.jags]{runjags::run.jags()}}}
}
\value{
A runjags object with some modifications:
\itemize{
\item the data slot is a in list.format, converted using \code{\link[=as_list_format]{as_list_format()}}
\item elements in the data slot have dimension names given by the input data frames
\item the slot XoccProcess is the process used to prepare the occupancy covariates (scaling and centering).
It can be applied using \code{\link[=apply.designmatprocess]{apply.designmatprocess()}}
\item the slot XobsProcess is the process used to prepare the detection covariates (scaling and centering).
It can be applied using \code{\link[=apply.designmatprocess]{apply.designmatprocess()}}
\item ModelSite slot is the list with values for each visit (row in detection covariates) giving the row of the ModelSite in the occupancy covariates.
\item species slot is the list of species names.
}
}
\description{
Fit detection occupancy models using runjags.
}
\section{Functions}{
\itemize{
\item \code{prep.data}: Given the input data parameters of run.detectionoccupancy prepare the data list for JAGS

\item \code{prep_new_data}: A short function that applies the prep.data function to new data, given an object created by run.detectionoccupancy
Xocc, yXobs, ModelSite must follow some rules as for run.detectionoccupancy, except yXobs may omit the species detections

\item \code{defaultinitsfunction}: Specifies the initial conditions for the MCMC chains.
This functions is called by \code{\link[=run.detectionoccupancy]{run.detectionoccupancy()}}
}}

