% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_function.R
\name{save_process}
\alias{save_process}
\title{Save and load a data preparation function}
\usage{
save_process(fun, checkwith, params = NULL)
}
\arguments{
\item{fun}{A data preparation function with first argument \code{indf} and other fixed parameters.}

\item{checkwith}{is a an object used only for checking that the function does not depend on items in the global environment}

\item{params}{A named list of any parameters that specify the transformation that fun performs.}
}
\description{
Save function without saving all of the global environment
}
