% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fit.R
\name{simulate.fit}
\alias{simulate.fit}
\title{Simulate observations from parameters of a fitted object.}
\usage{
\method{simulate}{fit}(fit, esttype = "median", conditionalLV = TRUE)
}
\arguments{
\item{fit}{A runjags fitted object created by \code{\link[=run.detectionoccupancy]{run.detectionoccupancy()}}}

\item{esttype}{Specifies parameter set to extract from fit, see \code{\link[=get_theta]{get_theta()}}}

\item{conditionalLV}{Logical. If TRUE, the simulation uses fitted LV values.}
}
\description{
Simulate observations from parameters of a fitted object.
}
\examples{
fit <- readRDS("./tmpdata/deto_wind.rds")
fit$data <- as.list.format(fit$data)
detected <- simulate.fit(fit, esttype = "median")

## Use to check residuals
fit_sim <- fit
fit_sim$data$y <- detected
source("./R/DS_residuals.R")
source("./R/DS_residuals_plots.R")
resid_det <- ds_detection_residuals.fit(fit_sim, type = 100, seed = 321)
resid_occ <- ds_occupancy_residuals.fit(fit_sim, type = "median", seed = 123)
## These residuals *should* be perfectly normally distributed for simulated data.
resid_occ <- ds_occupancy_residuals.fit(fit_sim, seed = 123)
plot_residuals_detection.fit(fit_sim, 
                   detectionresiduals = resid_det,
                   varidx = 2,
                   plotfunction = facet_covariate) +
coord_cartesian(ylim = c(-1, 1))
plot_residuals_detection.fit(fit_sim, 
                   detectionresiduals = resid_det,
                   varidx = 2) +
coord_cartesian(ylim = c(-1, 1))
}
