% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fit.R
\name{simulate_fit}
\alias{simulate_fit}
\title{Simulate observations from parameters of a fitted object.}
\usage{
simulate_fit(fit, esttype = "median", UseFittedLV = TRUE)
}
\arguments{
\item{fit}{A runjags fitted object created by \code{\link[=run.detectionoccupancy]{run.detectionoccupancy()}}}

\item{esttype}{Specifies parameter set to extract from fit, see \code{\link[=get_theta]{get_theta()}}}

\item{UseFittedLV}{Logical. If TRUE, the simulation uses fitted latent variable values.
If FALSE, latent variable values will be simulated for each ModelSite}
}
\description{
Simulate observations from parameters of a fitted object.
}
\examples{
fit <- readRDS("./tmpdata/deto_wind.rds")
fit$data <- as_list_format(fit$data)
detected <- simulate_fit(fit, esttype = "median", UseFittedLV = TRUE)

}
