% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVofrandomselectRV.R
\name{sumRV_margrow}
\alias{sumRV_margrow}
\title{For a list of random variables, return random variable that is a random selection one of the RVs}
\usage{
sumRV_margrow(pmat)
}
\arguments{
\item{rvs}{A list of Random Variable (distributions) of class 'RV'}
}
\description{
For a list of random variables, return random variable that is a random selection one of the RVs
}
\examples{
library(discreteRV)
X1 <- RV(c(1,0), c(.5,.5))
X2 <- RV(c(1,0), c(.5,.5))
X3 <- RV(c(1,-1), c(.5,.5))
randselRV(list(X1, X2, X3))
}
