% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_2_sp_points.R
\name{sws_sites_2_spdf}
\alias{sws_sites_2_spdf}
\alias{sws_sites_2_sf}
\title{Convert SWS Site Data in an SPDF}
\usage{
sws_sites_2_spdf(swssite_dataframe)

sws_sites_2_sf(swssite_dataframe)
}
\arguments{
\item{swssite_dataframe}{A data frame of the SWS site data}
}
\value{
An SpatialPointsDataFrame version of the SWS side data. Projection information is not included yet.
}
\description{
Reads in THE data frame of SWS site data and converts it into a Spatial Points Data Frame.
Projection information is not included yet.
}
\section{Functions}{
\itemize{
\item \code{sws_sites_2_sf}: Convert SWS Site Data in SF Object
}}

\examples{
sws_sites <- readRDS("./private/data/clean/sws_sites.rds")
sws_sites_spdf <- sws_sites_2_spdf(sws_sites)
sws_sites_sf <- sws_sites_2_sf(sws_sites)
}
