% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woody_vals_buffer.R
\name{woody_vals_buffer}
\alias{woody_vals_buffer}
\title{Extract mean woody cover within a buffer of a point}
\usage{
woody_vals_buffer(roi, pts, years, buffer)
}
\arguments{
\item{roi}{is an sf object.}

\item{pts}{is a set of sf points within roi}

\item{years}{Years of data to acquire}

\item{buffer}{is in METRES!}
}
\value{
each row is a point, each column is a date
}
\description{
Extract mean woody cover within a buffer of a point
}
