% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdesignmat.R
\name{prep.designmatprocess}
\alias{prep.designmatprocess}
\alias{apply.designmatprocess}
\alias{unstandardise.designmatprocess}
\title{Preprocessing Input Data}
\usage{
prep.designmatprocess(indata, fmla, version = 2, ...)

apply.designmatprocess(designmatprocess, indata)

unstandardise.designmatprocess(designmatprocess, indata)
}
\arguments{
\item{indata}{Input dataframe to be processed.}

\item{fmla}{A model formula (predictor side only).}

\item{designmatprocess}{Are instructions for preprocessing input data, created by \code{\link[=prep.designmatprocess]{prep.designmatprocess()}}}

\item{stoponhighcorrelation}{If TRUE the preparations for a design matrix will fail if correlations between covariates are higher than 0.75}

\item{data}{Dataframe to be processed.}
}
\value{
A special list containing parameters for applying a preprocessing step to data.

A design matrix.

The columns of indata before centering and scaling
}
\description{
Preprocessing Input Data
}
\details{
The input data is turned into a design matrix using \code{\link[stats:model.matrix]{stats::model.matrix()}}.
Each non-constant column is then centered and scaled.
}
\section{Functions}{
\itemize{
\item \code{prep.designmatprocess}: Calculates the parameters required to build a centred and scaled design matrix from input data.

\item \code{apply.designmatprocess}: Builds a centred and scaled design matrix from input data.

\item \code{unstandardise.designmatprocess}: Uncentres and unscales already standardised data.
}}

