% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks_of_rs_products.R
\name{brick_gpp}
\alias{brick_gpp}
\alias{brick_pg}
\alias{brick_ssoil}
\alias{brick_fmc}
\title{Extract Bricks of Remote Sensing Data}
\usage{
brick_gpp(spobj, years)

brick_pg(spobj, years)

brick_ssoil(spobj, years)

brick_fmc(spobj, years)
}
\arguments{
\item{spobj}{Spatial* object that informs extents of the raster to extract}

\item{years}{Years of data to extract}
}
\value{
A raster brick with extent equal or larger than \code{extent(spobj)}, snapped to the cells of the raster data. The crs of the GPP brick is WGS84.
}
\description{
Extract Bricks of Remote Sensing Data
}
\section{Functions}{
\itemize{
\item \code{brick_pg}: Extract brick of pg values

\item \code{brick_ssoil}: Extract brick of soil moisture values

\item \code{brick_fmc}: Extract brick of 8 day fmc values
}}

