% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugsdataconversion.R
\name{bugsvar2array}
\alias{bugsvar2array}
\alias{bugsvar2matrix}
\alias{matrix2bugsvar}
\title{Converting Parameter Values in BUGS Name and Format to Arrays and Matrices}
\usage{
bugsvar2array(values, varname, rowidx, colidx)

bugsvar2matrix(values, varname, rowidx, colidx)

matrix2bugsvar(theta, name)
}
\arguments{
\item{values}{is a list of values named according to the bugs variables names}

\item{varname}{is the desired variable name (e.g. 'u.b')}

\item{rowidx}{is a list of rows to extract, by number}

\item{colidx}{is a list of columns to extract, by number}

\item{theta}{is a parameter arrays}

\item{name}{parameter}
}
\value{
a named vector. Names are given by name and the index in the array
}
\description{
For converting values for array-valued parameters from the bugs variable format to an array
}
\section{Functions}{
\itemize{
\item \code{bugsvar2matrix}: For converting values for array-valued parameters from the bugs variable format to a matrix

\item \code{matrix2bugsvar}: Converts a matrix of parameter values to bugs variable format
}}

