% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpds_compare.R
\name{elpd_compare}
\alias{elpd_compare}
\title{Compare Sitewise Differences in Expected Log Posterior Density}
\usage{
elpd_compare(elpds, refname = NULL)
}
\arguments{
\item{elpds}{A data.frame or matrix of expected log posterior densities. Each row is ModelSite, each column is a Model. Typically the columns will have names.}

\item{refname}{The column name (i.e. Model) to compare all other models to. If NULL then uses the model with the highest sum of elpd.}
}
\description{
Compare Sitewise Differences in Expected Log Posterior Density
}
\details{
Very similar to loo::loo_compare(), but with fewer checks
}
