% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theta.R
\name{get_theta}
\alias{get_theta}
\title{A helper function to get a vector of parameters from a fitted object}
\usage{
get_theta(fit, type)
}
\arguments{
\item{fit}{fitted runjags object with summary included}

\item{type}{An integer will select the draw from the posterior
(e.g. type = 5 will return the vector of parameters in the 5th sample from the posterior, mcmc chains are concatenated)
A charactor vector can be used to the parameters based on summarary statistics like quantiles and moments.
Supported values so far "median" and "mean".
If \link{type} is "marginal" or "all" then all draws are returned.}
}
\description{
A helper function to get a vector of parameters from a fitted object
}
