% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tilecodes.R
\name{get_tilecodes}
\alias{get_tilecodes}
\title{Intersecting Albers Australian Tile Codes}
\usage{
get_tilecodes(spobj)
}
\arguments{
\item{spobj}{an sp object}
}
\description{
Given a spatial object returns the tile codes for Albers Tiles used by Geoscience Australia and others
}
