% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_onlineinfo.R
\name{ll2webdata}
\alias{ll2webdata}
\title{Download online data for spatial locations that are needed for particular models.}
\usage{
ll2webdata(coords, years, model = "7_4")
}
\arguments{
\item{coords}{An \link{sf} object with coordinates.}

\item{years}{Years of woody cover to use.}

\item{model}{A character code providing the model for which data to acquire}
}
\value{
A dataframe with columns of woody500m each year, and climatic information.
}
\description{
Downloads required data for spatial locations in order to run model.
}
\examples{
locations <- read.csv("./private/data/clean/site_locations_cleaned.csv")
locs <- sf::st_as_sf(locations, coords = c("MeanLON", "MeanLAT"), crs = 4326)
coords <- locs[ locs$StudyCode == "Nanangroe Natural Experiment", ]
fromwebdata <- ll2webdata(coords, 2018)
}
