% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_biodiversity.R
\name{pdetection_occupied.ModelSite.theta}
\alias{pdetection_occupied.ModelSite.theta}
\alias{poccupy.ModelSite.theta}
\title{Probability of detection and occupation for a given ModelSite}
\usage{
pdetection_occupied.ModelSite.theta(Xobs, v.b)

poccupy.ModelSite.theta(Xocc, u.b, lv.coef, LVvals)
}
\arguments{
\item{v.b}{Covariate loadings. Each row is a species, each column a detection covariate}

\item{u.b}{Covariate loadings for occupancy. Each row is a species, each column an occupancy covariate.}

\item{lv.coef}{Loadings for the latent variables. Each row is a species, each column corresponds to a LV.}
}
\value{
A matrix of detection probabilities for each row of Xobs and each species.
Rows in returned value correspond to rows in Xobs, columns correspond to species.

A matrix of occupancy probabilities. Each row corresponds to a row of LVvals, each column to a species.
}
\description{
Probability of detection and occupation for a given ModelSite
}
\section{Functions}{
\itemize{
\item \code{poccupy.ModelSite.theta}: The probability of occupation for given LV values.
}}

