% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_biodiversity.R
\name{expectedspeciesnum.ModelSite.theta}
\alias{expectedspeciesnum.ModelSite.theta}
\alias{predsumspecies}
\alias{predsumspecies_raw}
\alias{predsumspecies_newdata}
\title{Expected Biodiversity}
\usage{
expectedspeciesnum.ModelSite.theta(
  Xocc,
  Xobs = NULL,
  u.b,
  v.b = NULL,
  lv.coef = NULL,
  LVvals = NULL
)

predsumspecies(
  fit,
  chains = NULL,
  UseFittedLV = TRUE,
  nLVsim = 1000,
  cl = NULL
)

predsumspecies_raw(
  Xocc,
  Xobs,
  ModelSite,
  numspecies,
  nlv,
  draws,
  useLVindraws = TRUE,
  nLVsim = NULL,
  cl = NULL
)

predsumspecies_newdata(
  fit,
  Xocc,
  Xobs,
  ModelSiteVars,
  chains = NULL,
  nLVsim = 1000,
  cl = NULL
)
}
\arguments{
\item{Xocc}{A matrix of occupancy covariates, each row is a ModelSite}

\item{Xobs}{A matrix of detection covariates. Each row is a visit. The visited ModelSite (row of Xocc) is given by ModelSite}

\item{LVvals}{A matrix of LV values. Each column corresponds to a LV. To condition on specific LV values, provide a matrix of row 1.}

\item{chains}{The chains of MCMC to use. Default is all chains.}

\item{UseFittedLV}{If TRUE the fitted LV variables are used, if false then 1000 LV values are simulated.}

\item{nLVsim}{The number of simulated LV values if not using fitted LV values (only applies if  useLVindraws = FALSE).}

\item{cl}{A cluster object created by parallel::makeCluster. If NULL no cluster is used.}

\item{ModelSite}{is a list of integers giving the row in Xocc corresponding to a row in Xobs}

\item{draws}{A matrix of posterior parameter draws. Each row is a draw. Column names follow the BUGS naming convention}

\item{useLVindraws}{Use the LV values corresponding to each draw from within the \code{draws} object.
If FALSE nLVsim simulated LV values will be used for each draw.}

\item{theta}{A vector of model parameters, labelled according to the BUGS labelling convention seen in runjags}
}
\value{
A named vector of the expectation and variance of the numbers of species occupying the ModelSite and given parameter set.
If observational covariates are supplied then the expection and variance of numbers of species detected is also returned.

A matrix with each column a ModelSite.
Each row is labelled and corresponds to the predicted expection and variance of the number of species occupied or detected.
These expectations are with respect to the full posterior distribution of the model parameters, with the exception of the LV values which depends on UseFittedLV.

A matrix. Each column is a model site, each row is a different summary of the number of species.
There will be four rows: the expection and variance of the number of species occupied or detected.
}
\description{
The expected number of species occupying a ModelSite, or the expected number of species detected at a ModelSite.
}
\details{
No scaling or centering of Xocc or Xobs is performed by predsumspecies_raw
}
\section{Functions}{
\itemize{
\item \code{expectedspeciesnum.ModelSite.theta}: Computes expected numbers of species for a single parameter set and single ModelSite

\item \code{predsumspecies_raw}: Called by predsumspecies and predsumspecies_newdata.
Given ModelSite data and model information, compute expected and variance of the number of species.

\item \code{predsumspecies_newdata}: For new ModelSite occupancy covariates and detection covariates, predicted number of expected species
}}

\examples{
fit <- readRDS("./tmpdata/7_2_9_addyear_msnm_year_time_2lv.rds")
theta <- get_theta(fit, type = 1)
Xocc <- fit$data$Xocc[2, , drop = FALSE]
Xobs <- fit$data$Xobs[fit$data$ModelSite == 2, , drop = FALSE]
numspecies <- fit$data$n
lvsim <- matrix(rnorm(2 * 1), ncol = 2, nrow = 2) #dummy lvsim vars
lv.coef.bugs <- matrix2bugsvar(matrix(0, nrow = fit$data$n, ncol = 2), "lv.coef")
theta <- c(theta, lv.coef.bugs)

expectedspeciesnum.ModelSiteIdx(fit, 2, chains = NULL, LVvals = NULL)
Enumspec <- predsumspecies(fit, UseFittedLV = TRUE)


indata <- readRDS("./private/data/clean/7_2_10_input_data.rds")
predsumspecies_newdata(fit, Xocc <- indata$holdoutdata$Xocc, Xobs = indata$holdoutdata$yXobs, ModelSiteVars = "ModelSiteID", draws, cl = NULL)

}
