% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numspec_fulldistribution.R
\name{predsumspeciesRV}
\alias{predsumspeciesRV}
\alias{sumspeciesRV_raw}
\title{Compute predicted distributions of number species detected}
\usage{
predsumspeciesRV(
  fit,
  chains = NULL,
  UseFittedLV = TRUE,
  nLVsim = 100,
  type = "marginal",
  cl = NULL
)

sumspeciesRV_raw(
  Xocc,
  Xobs,
  ModelSite,
  numspecies,
  nlv,
  draws,
  useLVindraws = TRUE,
  nLVsim = NULL,
  cl = NULL
)
}
\arguments{
\item{fit}{A runjags fitted object}

\item{chains}{The chains of MCMC to use. Default is all chains.}

\item{UseFittedLV}{If TRUE the fitted LV variables are used, if false then 100 LV values are simulated.}

\item{type}{If "draws" then a list of distributions for each site and draw is returned
If "marginal" then a single distribution for each site is returned and this distribution is marginalised over the draws.}

\item{cl}{A cluster object created by parallel::makeCluster. If NULL no cluster is used.}

\item{Xocc}{A matrix of occupancy covariates. Must have a single row. Columns correspond to covariates.}

\item{Xobs}{A matrix of detection covariates, each row is a visit. If NULL then expected number of species in occupation is returned}

\item{theta}{A vector of model parameters, labelled according to the BUGS labelling convention seen in runjags}

\item{LVvals}{A matrix of LV values. Each column corresponds to a LV. To condition on specific LV values, provide a matrix of row 1.}
}
\description{
Uses \link{discreteRV} to compute distribution of number of species
}
\section{Functions}{
\itemize{
\item \code{sumspeciesRV_raw}: For raw model information. Returns distributions per draw.
}}

\examples{
fit <- readRDS("./tmpdata/7_2_9_addyear_msnm_year_time_2lv.rds")
theta <- get_theta(fit, type = "median")
u.b <- bugsvar2matrix(theta, "u.b", 1:fit$data$n, 1:ncol(fit$data$Xocc))
v.b <- bugsvar2matrix(theta, "v.b", 1:fit$data$n, 1:ncol(fit$data$Xobs))
lv.coef <- bugsvar2matrix(theta, "lv.coef", 1:fit$data$n, 1:fit$data$nlv)
nLVsim <- 1000
lvsim <- matrix(rnorm(fit$data$nlv * nLVsim), ncol = fit$data$nlv, nrow = nLVsim)
LVvals <- bugsvar2matrix(theta, "LV", 1:fit$data$n, 1:fit$data$nlv)
sumRV_margLV <- speciesnum.ModelSite.theta(Xocc = fit$data$Xocc[1, , drop = FALSE], u.b = u.b, lv.coef = lv.coef, LVvals = lvsim)
quantile(sumRV_margLV, probs = c(0.025, 0.5, 0.975))
sumRV_fitLV <- speciesnum.ModelSite.theta(Xocc = fit$data$Xocc[1, , drop = FALSE], u.b = u.b, lv.coef = lv.coef, LVvals = LVvals[1, , drop = FALSE])
quantile(sumRV_fitLV, probs = c(0.025, 0.5, 0.975))
full distribution of species number given fixed theta and unknown LV value (i.e. many simulated LV values). This is *within* model uncertainy to biodiversity

}
